refactor(infra): added infra layer for database and http services
In the database layer, the data mapper pattern and the repository pattern for the orm prisma are
used.
In the http layer the controllers will be created.
I redid the prisma schema, deleting the
old migrations and creating a model customer in the schema.

BREAKING CHANGE: The organization of the folders is now structured in a domain layer and an infra
layer, where the infra layer contains the application's core connection with the database and http
services